fix: remove more stuff
