+++ packages/strategy-tests/src/query.rs
@@ -1,5 +1,5 @@
 use crate::frequency::Frequency;
-use crate::strategy::StrategyRandomness;
+use crate::StrategyRandomness;
 use dapi_grpc::platform::v0::{
     get_identities_by_public_key_hashes_response, GetIdentitiesByPublicKeyHashesRequest,
     GetIdentitiesByPublicKeyHashesResponse, Proof,
